LayoutTests/media/modern-media-controls/layout-node/width-expected.txt

Testing the LayoutNode.width property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Checking default value
PASS node.width is 0
PASS node.element.style.width is ""

node.width = 10
PASS node.width is 10
PASS node.element.style.width is ""

node.width = 20

Layout was performed
PASS node.width is 20
PASS node.element.style.width is "20px"

node.width = 20
PASS scheduler.hasScheduledLayoutCallbacks is false

PASS successfullyParsed is true

TEST COMPLETE
